feat : added test to sport and rank CRUD

+++ src/rank/use-cases/rank.remove.ts
@@ -3,16 +3,20 @@ import { Repository } from 'typeorm';
 import { Rank } from '../entitites/rank.entity';
 import { Sport } from 'src/sport/entities/sport.entity';
 import { v4 } from 'uuid';
+import { NotFoundException } from '../exceptions/notFound.exception';
 
 @Injectable()
 export class RemoveRank {
-  constructor(
-    @Inject('RANK_REPOSITORY')
-    private rankRepository: Repository<Rank>,
-  ) {}
+    constructor(
+        @Inject('RANK_REPOSITORY')
+        private rankRepository: Repository<Rank>,
+    ) {}
 
-  async execute(id: string) {
-    const rankToRemove = await this.rankRepository.findOneBy({ id: id });
-    return await this.rankRepository.remove(rankToRemove);
-  }
+    async execute(id: string) {
+        const rankToRemove = await this.rankRepository.findOneBy({ id: id });
+        if (!rankToRemove) {
+            throw new NotFoundException();
+        }
+        return await this.rankRepository.remove(rankToRemove);
+    }
 }
